+++ Sources/BraintreePayPal/BTPayPalRequest.swift
@@ -181,7 +181,7 @@ import BraintreeCore
     public func parameters(
         with configuration: BTConfiguration,
         universalLink: URL? = nil,
-        fallbackUrlScheme: String? = nil,
+        fallbackURLScheme: String? = nil,
         isPayPalAppInstalled: Bool = false
     ) -> [String: Any] {
         var experienceProfile: [String: Any] = [:]
@@ -243,15 +243,12 @@ import BraintreeCore
         }
  
         if let universalLink, enablePayPalAppSwitch, isPayPalAppInstalled {
-            var appSwitchParameters: [String: Any] = [
+            let appSwitchParameters: [String: Any] = [
                 "launch_paypal_app": enablePayPalAppSwitch,
                 "os_version": UIDevice.current.systemVersion,
                 "os_type": UIDevice.current.systemName,
                 "merchant_app_return_url": universalLink.absoluteString
             ]
-            if let fallbackUrlScheme {
-                appSwitchParameters["fallback_url_scheme"] = fallbackUrlScheme
-            }
             
             return parameters.merging(appSwitchParameters) { $1 }
         }